Make separate module for DOM manipulation
